Made some styling for QuizScreen. It's not working too good.
This is more of a note if someone wants to work with this code. The buttons were working for correct Answer, but the wrong answers were set up as an array, so it won't work. Also, for some reason the back arrow next to App won't work, and the background is all orange. Everything is fine. Everything is fine.